+++ NEWS.md
@@ -30,6 +30,7 @@
 - Make `dynamic` a formal argument of `target()`.
 - Always lock/unlock the environment target by target, allowing informative error messages to appear more readily (#1062, @PedramNavid)
 - Automatically ignore `storr`s and decorated `storr`s (#1071).
+- Speed up memory management by avoiding a call to `setdiff()`.
 
 
 # Version 7.7.0